refactor(marketplace): move stepped buy dialog into strategy
This is so that the dialog and transaction code that the dialog is
executing can be extended for when a Safe multisig is selected in the
dApp.

The next commit will introduce the SafeBuyFractionalStrategy that
implements the different calls to the marketplace SDK.
