+++ marketplace/BuyFractionalStrategy.ts
@@ -1,0 +1,47 @@
+import {
+  HypercertExchangeClient,
+  WETHAbi,
+  addressesByNetwork,
+  utils,
+} from "@hypercerts-org/marketplace-sdk";
+import { readContract } from "viem/actions";
+import { UseWalletClientReturnType } from "wagmi";
+import type { AppRouterInstance } from "next/dist/shared/lib/app-router-context.shared-runtime";
+import { useStepProcessDialogContext } from "@/components/global/step-process-dialog";
+
+import { MarketplaceOrder } from "./types";
+
+export abstract class BuyFractionalStrategy {
+  constructor(
+    protected address: `0x${string}`,
+    protected chainId: number,
+    protected exchangeClient: HypercertExchangeClient,
+    protected dialogContext: ReturnType<typeof useStepProcessDialogContext>,
+    protected walletClient: UseWalletClientReturnType,
+    protected router: AppRouterInstance,
+  ) {}
+
+  abstract execute(params: {
+    order: MarketplaceOrder;
+    unitAmount: bigint;
+    pricePerUnit: string;
+    hypercertName?: string | null;
+    totalUnitsInHypercert?: bigint;
+  }): Promise<void>;
+
+  async getERC20Allowance(currency: `0x${string}`): Promise<bigint> {
+    const hypercertsExchangeAddress =
+      addressesByNetwork[utils.asDeployedChain(this.chainId)].EXCHANGE_V2;
+
+    if (!this.walletClient.data) {
+      return BigInt(0);
+    }
+
+    return (await readContract(this.walletClient.data, {
+      abi: WETHAbi,
+      address: currency as `0x${string}`,
+      functionName: "allowance",
+      args: [this.address, hypercertsExchangeAddress],
+    })) as bigint;
+  }
+}
